Working notification counter
