+++ app/styles/modules/_m-orange-bar.scss
@@ -133,3 +133,13 @@ ul .menu li.dropdown {
   flex-wrap: wrap;
   flex-direction: row;
 }
+
+button .orange-bar-notification-count {
+  position: absolute;
+  top: -10px;
+  right: -10px;
+  padding: 5px 10px;
+  border-radius: 50%;
+  background: #FFC861;
+  color: black;
+}
