--- app/templates/components/orange-bar-data-explorer-download-csv.hbs
@@ -1,1 +1,0 @@
-<a class='button tiny gray text-orange text-weight-normal' {{action 'exportTableToCSV'}}>{{fa-icon 'download'}} Download Data <small>(CSV)</small></a>
